+++ crates/env/src/engine/off_chain/tests.rs
@@ -39,13 +39,13 @@ fn topics_builder() -> Result<()> {
         // then
         assert_eq!(builder.topics.len(), 2);
 
-        let topics_len_compact = &scale::Compact(2u32);
-        let topics_len_encoded = scale::Encode::encode(&topics_len_compact);
         let output = TopicsBuilderBackend::<crate::DefaultEnvironment>::output(builder);
         #[rustfmt::skip]
-        let expected = vec![topics_len_encoded[0], 13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 17, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];
+        let expected = vec![
+            [13, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
+            [17, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]
+        ];
         assert_eq!(output, expected);
-
         Ok(())
     })
 }