+++ crates/ink/macro/src/lib.rs
@@ -18,6 +18,8 @@
 )]
 
 extern crate proc_macro;
+#[macro_use]
+extern crate ink_codegen;
 
 mod blake2b;
 mod chain_extension;
@@ -1389,8 +1391,8 @@ synstructure::decl_derive!(
     ///     b: [u8; 32],
     /// }
     ///
-    /// ink_env::emit_event::<DefaultEnvironment, _>(MyEvent { a: 42, b: [0x42; 32] });
-    /// ink_env::emit_event::<DefaultEnvironment, _>(MyAnonEvent { a: 42, b: [0x42; 32] });
+    /// ink_env::emit_event::<DefaultEnvironment, _, ink::abi::Ink>(&MyEvent { a: 42, b: [0x42; 32] });
+    /// ink_env::emit_event::<DefaultEnvironment, _, ink::abi::Ink>(&MyAnonEvent { a: 42, b: [0x42; 32] });
     /// ```
     ///
     /// # The Signature Topic