+++ integration-tests/all-abi/events/event-def-unused/src/lib.rs
@@ -1,0 +1,13 @@
+#![cfg_attr(not(feature = "std"), no_std, no_main)]
+
+#[ink::trait_definition]
+pub trait FlipperTrait {
+    #[ink(message)]
+    fn flip(&mut self);
+}
+
+#[ink::event]
+pub struct EventDefUnused {
+    #[ink(topic)]
+    pub hash: [u8; 32],
+}