+++ linting/extra/src/lib.rs
@@ -23,6 +23,7 @@ dylint_linting::dylint_library!();
 
 extern crate rustc_ast;
 extern crate rustc_errors;
+extern crate rustc_hash;
 extern crate rustc_hir;
 extern crate rustc_index;
 extern crate rustc_lint;